+++ test/unit/transpile.js
@@ -1,13 +1,18 @@
 'use strict'
 
+/*jshint expr: true*/
+
 const expect = require('chai').expect
 const execSync = require('child_process').execSync
 const path = require('path')
 
 const cleanScopeAndRequire = require('../support/cleanScopeAndRequire')
+const helperModule = require('../integration/helper')
 
 describe('transpile', function(){
   const env = process.env
+  // need our fixtures in tmp for these tests
+  helperModule.call(this)
 
   beforeEach(cleanScopeAndRequire)
 
@@ -32,21 +37,27 @@ describe('transpile', function(){
     env.OPAL_RUNTIME_PATH = path.resolve(__dirname, '../support/tweakedOpalRuntime.js')
   }
 
-  function doTranspile(code, options, filename, relativeFileName) {
+  function getCompiler(code, options, filename, relativeFileName) {
     const targetOptions = {
       sourceRoot: process.cwd(),
       filename: filename || 'foo.rb',
       file: relativeFileName || 'foo.rb'
     }
     Object.assign(targetOptions, options)
-    return transpile(code, targetOptions, wpContext).code
+    return transpile(code, targetOptions, wpContext)
+  }
+
+  function doTranspile(code, options, filename, relativeFileName) {
+    return getCompiler(code, options, filename, relativeFileName).code
   }
 
   it('compiles an endpoint', function() {
-    var result = doTranspile('HELLO=123')
+    const result = getCompiler('HELLO=123')
+    const code = result.code
 
-    expect(result).to.include('Opal.cdecl($scope, \'HELLO\', 123)')
-    expect(result).to.not.include('Opal.modules')
+    expect(code).to.include('Opal.cdecl($scope, \'HELLO\', 123)')
+    expect(code).to.not.include('Opal.modules')
+    expect(result.requireTrees).to.be.false
   })
 
   it('passes bundled opal runtime through', function() {
@@ -89,7 +100,7 @@ describe('transpile', function(){
   })
 
   it('can use a Bundler provided version of Opal', function () {
-    this.timeout(6000)
+    this.timeout(10000)
 
     env.OPAL_USE_BUNDLER = 'true'
 
@@ -115,11 +126,17 @@ describe('transpile', function(){
     expect(result).to.include('Opal.modules["addtl_stub"]')
   })
 
+  it('indicates whether require_trees were used', function() {
+    const result = getCompiler('require_tree "./tree"', {}, path.resolve(__dirname, '../fixtures/tree.rb'), 'tree.rb')
+
+    expect(result.requireTrees).to.be.true
+  })
+
   describe('webpack requires', function() {
     it('standard', function() {
       var result = doTranspile('require "another_dependency"')
 
-      expect(result).to.match(/require\('!!the_loader_path\?file=another_dependency&requirable=true!.*\/test\/fixtures\/another_dependency\.rb'\);/)
+      expect(result).to.match(/require\('!!the_loader_path\?file=another_dependency&requirable=true!.*\/tmp\/fixtures\/another_dependency\.rb'\);/)
     })
 
     context('for compiler js file', function () {
@@ -203,7 +220,7 @@ describe('transpile', function(){
     it('JS require', function() {
       var result = doTranspile('require "pure_js"')
 
-      expect(result).to.match(/require\('.*test\/fixtures\/pure_js.js'\);/)
+      expect(result).to.match(/require\('.*tmp\/fixtures\/pure_js.js'\);/)
     })
 
     it('require_tree', function() {